+++ faker/providers/bank/fa_IR/__init__.py
@@ -54,7 +54,3 @@ class Provider(BankProvider):
         "بانک پاسارگاد",
         "بانک مشترک ایران-ونزوئلا",
     )
-
-    def bank(self) -> str:
-        """Generate a bank name."""
-        return self.random_element(self.banks)