+++ faker/providers/bank/nl_BE/__init__.py
@@ -63,10 +63,6 @@ class Provider(BankProvider):
         "XXX",
     ]
 
-    def bank(self) -> str:
-        """Generate a bank name."""
-        return self.random_element(self.banks)
-
     def bban(self) -> str:
         """Generate a valid BBAN."""
         account_number = self._generate_account_number()