+++ faker/providers/bank/zh_CN/__init__.py
@@ -31,7 +31,3 @@ class Provider(BankProvider):
         "恒丰银行",
         "西安银行",
     )
-
-    def bank(self) -> str:
-        """Generate a bank name."""
-        return self.random_element(self.banks)