+++ tests/providers/test_automotive.py
@@ -184,6 +184,18 @@ class TestItIt(_SimpleAutomotiveTestMixin):
     license_plate_pattern: Pattern = re.compile(r"[A-Z]{2}\d{3}[A-Z]{2}")
 
 
+class TestJaJp(_SimpleAutomotiveTestMixin):
+    """Test ja_JP automotive provider methods"""
+
+    license_plate_pattern: Pattern = re.compile(
+        r"^(?:品川|足立|練馬|横浜|川崎|名古屋|大阪|神戸|福岡|札幌|尾張小牧|伊勢志摩) "
+        r"\d{2,3} "
+        r"(?:あ|い|う|え|か|き|く|け|こ|さ|す|せ|そ|た|ち|つ|て|と|な|に|ぬ|ね|の|は|ひ|ふ|ほ|"
+        r"ま|み|む|め|も|や|ゆ|よ|ら|り|る|れ|ろ|わ|を) "
+        r"(?:\d{2}-\d{2}|・{1,3}\d{1,3})$"
+    )
+
+
 class TestKoKr(_SimpleAutomotiveTestMixin):
     license_plate_pattern: Pattern = re.compile(
         r"^\d{2,3}[가나다라마거너더러머버서어저고노도로모보소오조구누두루무부수우주]\d{4}$"