+++ tests/providers/test_company.py
@@ -130,6 +130,30 @@ class TestFrFr:
 
             vat_number = faker.company_vat(siren="123 456 789")
             assert vat_number == "FR 32 123 456 789"
+    APE_GENERIC_PATTERN = re.compile(r"^\d{2}\.\d{2}[A-Z]$")
+    APE_2003_PATTERN = re.compile(r"^\d{2}\.\d{2}[A-FZ]$")
+    APE_2025_PATTERN = re.compile(r"^\d{2}\.\d{2}[YGHJKL]$")
+
+    def test_ape_code(self, faker, num_samples):
+        for _ in range(num_samples):
+            # default version is "naf-2003"
+            code = faker.ape_code()
+            assert isinstance(code, str)
+            assert self.APE_2003_PATTERN.fullmatch(code), f"Invalid NAF 2003 APE code format: {code}"
+            # version naf-2003
+            code = faker.ape_code(version="naf-2003")
+            assert isinstance(code, str)
+            assert self.APE_2003_PATTERN.fullmatch(code), f"Invalid NAF 2003 APE code format: {code}"
+            # version naf-2025
+            code = faker.ape_code(version="naf-2025")
+            assert isinstance(code, str)
+            assert self.APE_2025_PATTERN.fullmatch(code), f"Invalid NAF 2025 APE code format: {code}"
+            # Possibly invalid numbers
+            code = faker.ape_code(version=None)
+            assert isinstance(code, str)
+            assert self.APE_GENERIC_PATTERN.fullmatch(code), f"Invalid APE code format: {code}"
+        with pytest.raises(ValueError):
+            faker.ape_code(version="naf-1984")
 
 
 class TestHyAm: